добавляет обработчик событий на закрытие окна отправки формы
