+++ js/main.js
@@ -1,5 +1,4 @@
-import {setAddress, setAdFormSubmit} from './form.js';
-import {getAdFormDisabled, getAdFormActive} from './ad-form.js';
+import {setAddress, setAdFormSubmit, getAdFormDisabled, getAdFormActive, resetForm} from './ad-form.js';
 import {initMap, setAdPins, setOnMapLoad, setOnMainPinMove} from './map.js';
 import {createSlider, setOnSliderUpdate} from './slider.js';
 import {getData} from './api.js';
@@ -24,4 +23,4 @@ getAdFormDisabled();
 initMap(START_COORDINATE);
 createSlider();
 setOnSliderUpdate();
-setAdFormSubmit();
+setAdFormSubmit(resetForm);
